src/requirements.txt
strsim
imdbpy
click
similarity
